2X (7s) speedup in the bmalloc clean build https://bugs.webkit.org/show_bug.cgi?id=312072 rdar://174582619
Reviewed by Richard Robinson.

Mostly changes to improve build parallelism.

* Source/bmalloc/CMakeLists.txt:
* Source/bmalloc/Configurations/bmalloc.xcconfig:
* Source/bmalloc/Configurations/libpas.xcconfig: Removed. Merged the bmalloc
and libpas targets. There's no logical separation anymore, and the two targets
are merged in the CMake build already.

Separate targets reduced parallelism because the second target waited for the
first.

* Source/bmalloc/bmalloc.swift: Removed. Don't build an unused Swift module.
It ends up precompiling many transitive dependecies -- but in a way that the
rest of the build does not currently reuse.

* Source/bmalloc/bmalloc.xcodeproj/project.pbxproj: Ditto.

* Source/bmalloc/bmalloc/TZoneHeap.cpp: Only #include bmalloc_heap_iso_internal.h
because it's much smaller.

* Source/bmalloc/libpas/libpas.xcodeproj/project.pbxproj:
* Source/bmalloc/libpas/src/libpas/bmalloc_heap.c:
* Source/bmalloc/libpas/src/libpas/bmalloc_heap.h:
* Source/bmalloc/libpas/src/libpas/bmalloc_heap_flex.c: Added.
* Source/bmalloc/libpas/src/libpas/bmalloc_heap_flex_internal.h: Added.
* Source/bmalloc/libpas/src/libpas/bmalloc_heap_inlines.h:
* Source/bmalloc/libpas/src/libpas/bmalloc_heap_iso.c: Added.
* Source/bmalloc/libpas/src/libpas/bmalloc_heap_iso_internal.h: Renamed from Source/bmalloc/libpas/src/libpas/bmalloc_heap_internal.h.
* Source/bmalloc/libpas/src/libpas/bmalloc_heap_utils.c: Added.
* Source/bmalloc/libpas/src/libpas/bmalloc_heap_utils.h: Added. Split bmalloc_heap.c
into separate files because it was enormous, which prevented parallelism. I only
split out files that would not require including bmalloc_heap_inlines.h, since
including that big header twice would outweigh the optimization.

Canonical link: https://commits.webkit.org/311034@main
