+++ Source/bmalloc/Configurations/bmalloc.xcconfig
@@ -27,6 +27,7 @@ INSTALLHDRS_SCRIPT_PHASE = YES;
 PRIVATE_HEADERS_FOLDER_PATH = $(BMALLOC_INSTALL_PATH_PREFIX)$(WK_LIBRARY_HEADERS_FOLDER_PATH)/bmalloc;
 PRODUCT_NAME = bmalloc;
 
+GCC_PREPROCESSOR_DEFINITIONS = $(inherited) PAS_BMALLOC_HIDDEN;
 GCC_SYMBOLS_PRIVATE_EXTERN = YES;
 STRIP_INSTALLED_PRODUCT = NO;
