+++ Source/bmalloc/bmalloc/TZoneHeap.cpp
@@ -30,7 +30,7 @@
 
 #include "TZoneHeapManager.h"
 #include "bmalloc.h"
-#include "bmalloc_heap_internal.h"
+#include "bmalloc_heap_iso_internal.h"
 #include "bmalloc_heap_ref.h"
 
 #if !BUSE(LIBPAS)
