+++ Source/bmalloc/libpas/src/libpas/bmalloc_heap.h
@@ -27,6 +27,7 @@
 #define BMALLOC_HEAP_H
 
 #include "bmalloc_heap_ref.h"
+#include "bmalloc_heap_utils.h"
 #include "pas_allocation_mode.h"
 #include "pas_primitive_heap_ref.h"
 #include "pas_reallocate_free_mode.h"
@@ -155,14 +156,6 @@ PAS_API void* bmalloc_reallocate_auxiliary(void* old_ptr,
 PAS_API pas_heap* bmalloc_auxiliary_heap_ref_get_heap(pas_primitive_heap_ref* heap_ref);
 
 PAS_API void bmalloc_deallocate(void*);
-
-PAS_API pas_heap* bmalloc_force_auxiliary_heap_into_reserved_memory(pas_primitive_heap_ref* heap_ref,
-                                                                    uintptr_t begin,
-                                                                    uintptr_t end);
-
-PAS_BAPI size_t bmalloc_heap_ref_get_type_size(pas_heap_ref* heap_ref);
-PAS_API pas_heap* bmalloc_get_heap(void* ptr);
-PAS_BAPI size_t bmalloc_get_allocation_size(void* ptr);
 
 PAS_END_EXTERN_C;
