+++ Source/bmalloc/libpas/src/libpas/bmalloc_heap_iso_internal.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2024 Apple Inc. All rights reserved.
+ * Copyright (c) 2024-2026 Apple Inc. All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
@@ -20,19 +20,45 @@
  * PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
  * OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE. 
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#ifndef BMALLOC_HEAP_INTERNAL_H
-#define BMALLOC_HEAP_INTERNAL_H
+#ifndef BMALLOC_HEAP_ISO_INTERNAL_H
+#define BMALLOC_HEAP_ISO_INTERNAL_H
 
-#include "pas_platform.h"
+#include "pas_config.h"
+
+#if LIBPAS_ENABLED
+
+PAS_IGNORE_WARNINGS_BEGIN("missing-field-initializers")
 
 #include "bmalloc_heap.h"
+#include "bmalloc_heap_config.h"
+#include "bmalloc_heap_innards.h"
+#include "pas_try_allocate.h"
+#include "pas_try_allocate_array.h"
+#include "pas_try_reallocate.h"
 
 #if PAS_ENABLE_BMALLOC
 
 PAS_BEGIN_EXTERN_C;
+
+PAS_CREATE_TRY_ALLOCATE(
+    bmalloc_try_iso_allocate_impl,
+    BMALLOC_HEAP_CONFIG,
+    &bmalloc_typed_runtime_config.base,
+    &bmalloc_allocator_counts,
+    pas_allocation_result_identity);
+
+PAS_CREATE_TRY_ALLOCATE(
+    bmalloc_iso_allocate_impl,
+    BMALLOC_HEAP_CONFIG,
+    &bmalloc_typed_runtime_config.base,
+    &bmalloc_allocator_counts,
+    pas_allocation_result_crash_on_error);
+
+PAS_API void* bmalloc_try_iso_allocate_casual(pas_heap_ref* heap_ref, pas_allocation_mode allocation_mode);
+PAS_API void* bmalloc_iso_allocate_casual(pas_heap_ref* heap_ref, pas_allocation_mode allocation_mode);
 
 PAS_CREATE_TRY_ALLOCATE_ARRAY(
     bmalloc_try_iso_allocate_array_impl,
@@ -212,113 +238,11 @@ static PAS_ALWAYS_INLINE void* bmalloc_iso_reallocate_array_by_count_inline(
         pas_reallocate_free_if_successful).begin;
 }
 
-PAS_CREATE_TRY_ALLOCATE_PRIMITIVE(
-    bmalloc_try_allocate_flex_impl,
-    BMALLOC_HEAP_CONFIG,
-    &bmalloc_flex_runtime_config.base,
-    &bmalloc_allocator_counts,
-    pas_allocation_result_identity);
-
-PAS_CREATE_TRY_ALLOCATE_PRIMITIVE(
-    bmalloc_allocate_flex_impl,
-    BMALLOC_HEAP_CONFIG,
-    &bmalloc_flex_runtime_config.base,
-    &bmalloc_allocator_counts,
-    pas_allocation_result_crash_on_error);
-
-PAS_API void* bmalloc_try_allocate_flex_with_alignment_casual(
-    pas_primitive_heap_ref* heap_ref, size_t size, size_t alignment, pas_allocation_mode allocation_mode);
-PAS_API void* bmalloc_allocate_flex_with_alignment_casual(
-    pas_primitive_heap_ref* heap_ref, size_t size, size_t alignment, pas_allocation_mode allocation_mode);
-
-static PAS_ALWAYS_INLINE void* bmalloc_try_allocate_flex_inline(
-    pas_primitive_heap_ref* heap_ref, size_t size, pas_allocation_mode allocation_mode)
-{
-    pas_allocation_result result;
-    result = bmalloc_try_allocate_flex_impl_inline_only(heap_ref, size, 1, allocation_mode);
-    if (PAS_LIKELY(result.did_succeed))
-        return (void*)result.begin;
-    return bmalloc_try_allocate_flex_with_alignment_casual(heap_ref, size, 1, allocation_mode);
-}
-
-static PAS_ALWAYS_INLINE void* bmalloc_allocate_flex_inline(
-    pas_primitive_heap_ref* heap_ref, size_t size, pas_allocation_mode allocation_mode)
-{
-    pas_allocation_result result;
-    result = bmalloc_allocate_flex_impl_inline_only(heap_ref, size, 1, allocation_mode);
-    if (PAS_LIKELY(result.did_succeed))
-        return (void*)result.begin;
-    return bmalloc_allocate_flex_with_alignment_casual(heap_ref, size, 1, allocation_mode);
-}
-
-static PAS_ALWAYS_INLINE void* bmalloc_try_allocate_zeroed_flex_inline(
-    pas_primitive_heap_ref* heap_ref, size_t size, pas_allocation_mode allocation_mode)
-{
-    return (void*)pas_allocation_result_zero(
-        bmalloc_try_allocate_flex_impl(heap_ref, size, 1, allocation_mode),
-        size).begin;
-}
-
-static PAS_ALWAYS_INLINE void* bmalloc_allocate_zeroed_flex_inline(
-    pas_primitive_heap_ref* heap_ref, size_t size, pas_allocation_mode allocation_mode)
-{
-    return (void*)pas_allocation_result_zero(
-        bmalloc_allocate_flex_impl(heap_ref, size, 1, allocation_mode),
-        size).begin;
-}
-
-static PAS_ALWAYS_INLINE void* bmalloc_try_allocate_flex_with_alignment_inline(
-    pas_primitive_heap_ref* heap_ref, size_t size, size_t alignment, pas_allocation_mode allocation_mode)
-{
-    pas_allocation_result result;
-    result = bmalloc_try_allocate_flex_impl_inline_only(heap_ref, size, alignment, allocation_mode);
-    if (PAS_LIKELY(result.did_succeed))
-        return (void*)result.begin;
-    return bmalloc_try_allocate_flex_with_alignment_casual(heap_ref, size, alignment, allocation_mode);
-}
-
-static PAS_ALWAYS_INLINE void* bmalloc_allocate_flex_with_alignment_inline(
-    pas_primitive_heap_ref* heap_ref, size_t size, size_t alignment, pas_allocation_mode allocation_mode)
-{
-    pas_allocation_result result;
-    result = bmalloc_allocate_flex_impl_inline_only(heap_ref, size, alignment, allocation_mode);
-    if (PAS_LIKELY(result.did_succeed))
-        return (void*)result.begin;
-    return bmalloc_allocate_flex_with_alignment_casual(heap_ref, size, alignment, allocation_mode);
-}
-
-static PAS_ALWAYS_INLINE void* bmalloc_try_reallocate_flex_inline(
-    pas_primitive_heap_ref* heap_ref, void* old_ptr, size_t new_size, pas_allocation_mode allocation_mode)
-{
-    return (void*)pas_try_reallocate_primitive(
-        old_ptr,
-        heap_ref,
-        new_size,
-        allocation_mode,
-        BMALLOC_HEAP_CONFIG,
-        bmalloc_try_allocate_flex_impl_for_realloc,
-        &bmalloc_flex_runtime_config.base,
-        pas_reallocate_disallow_heap_teleport,
-        pas_reallocate_free_if_successful).begin;
-}
-
-static PAS_ALWAYS_INLINE void* bmalloc_reallocate_flex_inline(
-    pas_primitive_heap_ref* heap_ref, void* old_ptr, size_t new_size, pas_allocation_mode allocation_mode)
-{
-    return (void*)pas_try_reallocate_primitive(
-        old_ptr,
-        heap_ref,
-        new_size,
-        allocation_mode,
-        BMALLOC_HEAP_CONFIG,
-        bmalloc_allocate_flex_impl_for_realloc,
-        &bmalloc_flex_runtime_config.base,
-        pas_reallocate_disallow_heap_teleport,
-        pas_reallocate_free_if_successful).begin;
-}
-
 PAS_END_EXTERN_C;
 
 #endif /* PAS_ENABLE_BMALLOC */
 
-#endif /* BMALLOC_HEAP_INTERNAL_H */
+PAS_IGNORE_WARNINGS_END
+
+#endif /* LIBPAS_ENABLED */
+#endif /* BMALLOC_HEAP_ISO_INTERNAL_H */
